feat(exception): implementa o tratamento de exceções da API
Cria as classes de exceção personalizadas ResourceNotFoundException
e DatabaseException para representar erros específicos da camada de serviço.

Adiciona um ControllerAdvice para capturar essas exceções de forma
global e retornar respostas HTTP adequadas (404, 400, etc.).

+++ src/main/java/com/aprendendospring/B/A/M/Commerce/services/ProductService.java
@@ -3,14 +3,20 @@ package com.aprendendospring.B.A.M.Commerce.services;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.aprendendospring.B.A.M.Commerce.dto.ProductDTO;
 import com.aprendendospring.B.A.M.Commerce.entities.Product;
 import com.aprendendospring.B.A.M.Commerce.repositories.ProductRepository;
+import com.aprendendospring.B.A.M.Commerce.services.exceptions.DatabaseException;
+import com.aprendendospring.B.A.M.Commerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -21,7 +27,8 @@ public class ProductService {
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
 		Optional<Product> result = repository.findById(id);
-		Product product = result.get();
+		Product product = result.orElseThrow(
+				() -> new ResourceNotFoundException("Recurso não encontrado"));
 		ProductDTO dto = new ProductDTO(product);
 		return dto;
 	}
@@ -42,15 +49,30 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
+		
+		try {
 		Product entity = repository.getReferenceById(id);
 		copyDtoToEntity(dto, entity);
 		entity = repository.save(entity);
 		return new ProductDTO(entity);
+		}
+		
+		catch(EntityNotFoundException e) {
+			throw new  ResourceNotFoundException("Recurso nao encontrado");
+		}
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+	        	repository.deleteById(id);    		
+		}
+	    	catch (DataIntegrityViolationException e) {
+	        	throw new DatabaseException("Falha de integridade referencial");
+	   	}
 	}
 
 	private void copyDtoToEntity(ProductDTO dto, Product entity) {
